requirements.txt
qrcode==7.4.2
Pillow==10.0.0
opencv-python==4.9.0.80

